Start progress subshell after FIFO reader is ready
Also:

- Shorten the sleep to 0.002 seconds to make the initial progress bar fill
  snappier.
- Use `stdbuf -oL` to apply line-buffered output to the FIFO pipe.
  This avoids some race conditions, and overall makes the progress bar fill
  smoother.

Two asynchronous subshell processes are started here:

- `for` loop writing to FIFO (for ... echo ...)
- `yad` progress dialog reading percentage from FIFO

Originally the `for` loop subshell running `echo "$i"` to the named pipe (FIFO)
would start _slightly before_ the yad dialog process was started.  This caused a
short time where writes to the pipe would block, before `yad` process could
begin reading from the FIFO. As a result, the `yad` process immediately reads a
few newline (`\n`) separated numeric percentage values from the FIFO pipe once
it was up and ready to read.  This caused the progress bar to appear to
jumpstart from a non-zero value.

Additionally, the inherent behavior of FIFO & PIPEs on Linux creates a situation
where these progress bar percentage values were subjected to buffered IO.  By
default, for a PIPE not connected to a tty (`isatty()` returns `0`), Linux
kernel will use buffered IO with a size of 4096 bytes.  For asynchronous
background processes, this is a **good thing**, because without a known buffer
size, writes cannot be assured to be atomic.  IEEE Std 1003.1-2001 requires
that writes of `PIPE_BUF` or fewer bytes shall be atomic.  Atomic writes ensure
that the amount of data written in one operation is not interleaved with data
from any other process.  This could otherwise happen when multiple writers send
data to a single reader, as in our modified `rofi-lpass` code.  For writes To be
performed atomically, applications must know how large the expected write
request is.  This value is `PIPE_BUF` on Linux/POSIX.1 (see `man 7 pipe`).

However, due to the buffered nature of PIPE IO, some race conditions surfaced
where the final `99` and `100` values may not have been read by `yad` by the
time that it was going to be killed.  This caused the final `echo '99'` and
`echo '100'` to sometimes block on FIFO write indefinitely, because there was no
longer a process reading from the FIFO pipe.

We will address this fully in later commits.