Refactor progress updates into function
Also, wait for lpass piped to wl-copy to both exit. This helps to handle the
case where `lpass` is waiting for user input (e.g. Master Password).  In that
case, the progress dialog will pause at 99% while the user interacts with
the `lpass` dialog window(s).  The final 1% will be added after `wl-copy` forks
off its child process to host the clipboard/selection offer via Wayland
protocols.

The `yad` progress dialog will be launched in that case, but simply wait until
the `wl-copy` has finished its work, which will generally guarantee that the
progress "security screen" dialog will still be opened and focused when the
actual clipboard event happens.  This is important because `wl-paste --watch`
will run the `cliphist` wrapper at that moment to try and store the value in
clipboard history.  After that happens, it's safe to send the final `echo '100'`
to the `yad` progress dialog window, because `cliphist` wrapper script should
have been able to see the focused `rofi-lpass:` dialog window title, and set
`CLIPBOARD_STATE=sensitive` by that time for the real `cliphist` binary to see
it.

All of this is a rather silly hack, which is still unfortunately necessary
because the Wayland clipboard (a.k.a. "selection") API protocols have no way of
signaling sensitive data at copy-time.  The Wayland protocols only allow for the
MIME type to be set on a piece of selection data offered.  Ideally, this could
be used to identify which pieces of data are potentially sensitive, and then
clipboard clients could handle things appropriately, such as to avoid storing
any sensitive data to clipboard history, adding some kind of configured
time-to-live before the data is removed from the clipboard or primary selection,
or other kinds of desired behavior.

Perhaps the time has come to standardize on some new MIME type such as
`text/sensitive`, or even the `x-kde-passwordManagerHint` from KDE & KeePassXC?

Either way, this would require changes to `wl-copy` to support sending one of
these MIME types at copy-time, and changes to `wl-paste --watch` to set
`CLIPBOARD_STATE=sensitive` when encountering data with one of these MIME types.