Code cleanup & remove more debugging output
Note: The `yad` options for closing window early
needed to be removed. This included the flags:

* `--close-on-unfocus`
* `--timeout=15`
* `--timeout-indicator=top`

If the window is closed early, the FIFO read end is closed and the
script will hang indefinitely.  This is because the writes to the FIFO
are blocking until a read is performed.  If the read end is closed, the
write will block indefinitely.

So, we must only rely on the last `100` progress update to close the window.