Watch via wl-paste until clipboard item is seen
The problem with the code before was due to timing of the cliphist command
and "security screen" dialog, which would not stay up long enough for the
cliphist wrapper script to see the clipboard item. This change adds a
background coprocess to watch for the clipboard item, verified via CRC hash
because the next immediate clipboard item is not always ours.  For example,
it could be the previous clipboard item that was not cleared yet.

When the expected CRC matches the clipboard item in `finalize-progress-dialog`,
we can at least be assured that `cliphist` also has just seen it at the same
time.  Further work will be needed to ensure that the cliphist wrapper script
has seen the security dialog and handled setting `CLIPBOARD_STATE=sensitive`
before the clipboard item is sent to the real `cliphist store` command.

This initial version sends only the CRC hash from mid-PIPE stream via `tee`, so
that the clipboard item is not sent in the clear.  It uses a named pipe in SHM
to communicate the CRC hash from the middle of the pipe stream to `wl-copy` out
of band to the main script process.  Note that SHM is used simply to store the
named PIPE path in the in-memory SHM virtual filesystem, not to store the CRC.
The Linux kernel will handle the IPC between the two processes using a FIFO
PIPE, which is a buffer in the kernel memory space.  The named PIPE is created
by the main script process and is used to communicate the CRC hash from the
middle of the pipe stream to the `wl-paste --watch` process. This is necessary
because the STDOUT of the pipeline was eaten by the `wl-copy` command, and we
need some way to pass the expected value to our `finalize-progress-dialog`
coprocess which will be launched by `wl-paste --watch` command with commandline
args in the clear.

The `finalize-progress-dialog` will be responsible for sending the final 100%
complete message to the security dialog process, `yad`, before cleaning up the
resources and exiting.

Further work must still be done to synchronize timing with the `cliphist`
wrapper script.  We will attempt to watch for that process to exit before
sending the final 100% complete message to the security dialog process.