Fix bug caused by shellcheck SC2086 auto-format on save
As it turns out, the reason that `cat -` and `cat <&0` were returning an empty
string was because on the selected line, `$PASTE_ONCE` was previously being
quoted as: `"$PASTE_ONCE"`.  This was happening automatically every time I
saved the file due to the `shfmt` and `shellcheck` VSCode extensions.
Apparently, when passing `"$PASTE_ONCE"` with the variable unset, it resulted in
the call to: `wl-copy ''`, which made `wl-copy` ignore STDIN and instad send an
empty string `''` to the clipboard!

Such a crazy mistake, and the code editor automatically made this change behind
my back, thanks to `shellcheck`'s SC2086 rule.