Fix: SC2030, SC2031 - subshell & global vars
The `SC2030` and `SC2031` warnings are due to the script using a
subshell, and trying to modify global variables within the subshell. The
script is using a subshell to run a coprocess in the background, and the
coprocess needs to access the global variables in the `trap` handler for
the `fork()`-ed subshell process.  Note that in this case, the subshell
is not an entirely separate and "fresh" `bash` interpreter, as in
`fork()` + `execve()` in a separate script or separate invocation.

The slightly confusing part here is that some of our functions are
designed to be re-entrant via a separate shell invocation, which is
_launched_ by either `rofi` (relaunched via its plugin API), or by the
`wl-paste --watch` command. The `rofi-lpass` script is designed to be
launched in multiple ways, and in some of these modes it is indeed an
entirely separate co-process.  However, within certain modes, some
functions call subshells which are disowned from the parent shell, and
these subshells are not entirely separate `bash` interpreters because
they have a copy of the shell environment as in `fork()`.

**Note:** There are some subtleties regarding `trap` inheritance and
other behaviors (e.g. `$$` versus `$BASHPID` will return different
values within a subshell, according to POSIX standard and Bash-specific
behavior).

The script within `show-progress-dialog()` was trying to set the global
variables within the subshell, but the shellcheck linter is warning that
the global variables are not being set in the parent shell and therefore
would have different values.

In our case, we don't really care because effectively the subshell is
using these for the `trap` handler, `kill-progress-dialog()`, and the
`tail -f` command is also using the local variable. So we could safely
ignore the shellcheck warnings in this case.  However, it is better to
set the global variables in the parent shell and pass them to the child.
This allows us to ensure that `PROGRESS_FIFO` global variable is set in
the same place where it is created, although we do try to pass it to
most of the various functions that need it as an argument.  This is a
bit of a trade-off between readability and maintainability, and the need
to avoid shellcheck warnings.  In any case, we should try to avoid using
a global variable in this way, except for the special case of `trap`.
This also avoids needing to place `shellcheck` comment directives all
over within the script.

The distictions and nuances between `fork()`-ed subshell(s) and
separated shell process invocations (`fork()` + `execve()`) matter here,
and can be quite esoteric.  In fact, even the venerable [Advanced
Bash-Scripting Guide][1] gets the definition of a subshell slightly
wrong!  This has been the cause of a lot of confusion regarding the
semantics of `()` subshells, global variables, `trap` handlers, and even
the POSIX `$$` shell-PID variable behavior.  Some of these things are
relevant to this script, so I'm adding this discussion here along with
some hopefully helpful references to clear up the details.

References:

  - https://unix.stackexchange.com/a/138498/7688
  - https://stackoverflow.com/a/51903792/645491
  - https://www.shellcheck.net/wiki/SC2030
  - https://www.shellcheck.net/wiki/SC2031

[1]: https://unix.stackexchange.com/q/138463/7688#comment776091_138498