Implement timeout to wait for cliphist store
This commit adds a timeout to the `inotifywait` command that waits for
the `cliphist store` process to end. If the `cliphist store` process
does not end within the timeout, the script will clean up the dialog,
kill the parent `wl-paste --watch` process, and return an error code to
indicate that the process did not end as expected.

The timeout is set to 10 seconds, which should be more than enough time.
If the process is not found within 10 seconds, it is likely that the
finalize-progress-dialog function started too late and has missed the
critical moment when it was running.

This change does not fix the problem of never seeing the correct
cliboard item CRC checksum, but it does help to ensure that the script
does not hang indefinitely if the `cliphist store` process is not found
as expected.