Fix shell floating point arithmetic error: use bc
Use `bc` to calculate the number of iterations to wait in the loop,
instead of using shell arithmetic, which rounds down and results in
floating point and syntax errors.  We may need microsecond precision in
this function, so we will need to use `bc` to do calculations.

Note: The floating point sleep interval is only supported in GNU sleep,
which supports floating point numbers.  The default sleep command on
some systems does not support floating point numbers.  Yet, we already
depend on GNU `stdbuf` from coreutils, so we should be fine because the
target system for this project is already GNU/Linux.