Fix timing issue with wl-paste --watch scripts
When copying sensitive data, the `rofi-lpass` script uses a pipeline to
copy the data. This pipeline includes a `wl-copy` command to copy the
data to the Wayland clipboard.  The timing of that command is critical
to ensure that the clipboard watcher script (`wl-paste --watch`) is
triggered AFTER the `rofi-lpass` script ends and the rofi window closes.

A few critical things need to happen with the right timing:

- The progress dialog, `yad`, must be visible and focused at the same
  time that the `cliphist store` process is launched by `wl-paste
  --watch`, which happens outside the `rofi-lpass` script, but is linked
  directly to when we launch `wl-copy`.  It is guaranteed that `cliphist
  store` will run some time in millisecods after `wl-copy` receives data
  and places it into the system clipboard.
- The initial `copy-account-field` code path must run as fast as
  possible, because the `rofi` dialog is running in Gnome + Wayland
  Layer Shell, and therefore steals focus from the progress dialog
  window when it is active.  We are guaranteed that the `rofi` dialog
  will close some momentary time after our `copy-account-field` script
  code path and shell process ends.  This also happens within
  milliseconds.
- The final call to `finalize-progress-dialog` is also guaranteed to
  happen after we call `wl-copy`, in the same way that `cliphist store`
  is guaranteed to run.  It also should run within milliseconds of the
  clipboard item becoming available in the system clipboard.

During testing, a screen recording video was made to gather information
about the timing of the above events.  Some observations about timing
were collected and timing data at critical points were analyzed. Viewing
the video in slow motion allowed me to gather these data points. Most of
the timing data from our script in trace mode can be trusted to
nanosecond precision, because all times are in seconds since the Unix
Epoch, also known as Unix timestamps, and all have 9 decimal points.

The main current problem observed during testing was that two critical
events happened at inopportune times:

1. The `yad` dialog appeared while `rofi` was still showing.  Normally,
   we want the dialog to appear as soon as possible, so its initial
   timing is ok. However, `rofi` was taking a bit too long to disappear
   because our `copy-account-field` function was still running at this
   time.
2. The `wl-copy` process is started within a pipeline in our
   `copy-account-field` function, which was originally much too early.
   As stated before, launching this command triggers the two other
   commands: `cliphist store` wrapper, and the
   `finalize-progress-dialog` function call, which is in our script.  We
   needed a way to delay the `wl-copy` command while still maintaining
   the previous constraints regarding the `cksum` FIFO, security of the
   clipboard data, and pipeline which it sits at the end of.

This change refactors the `wl-copy` command such that it can run
asynchronously, and at a much later time than the `lpass show` and
`cksum` commands in the pipeline.  Remember, it is critical that
`wl-copy` runs later, measured approximately between 367.467521 to
559.041926 milliseconds later than it originally did at the end of the
pipeline. Also remember that it's critically important that we have the
CRC checksum value from `cksum` command as soon as possible, which we
read from the `$cksum_pipe`.  We need this piece of data so we can
launch `wl-paste --watch` to call our `finalize-progress-dialog`
function using this `$pw_cksum` variable as a parameter. This has to
happen as quickly as possible so the first `copy-account-field` run of
our script completes as fast as possible and closes the `rofi` dialog.

So, we had a bit of a chicken and egg problem:

- We need the `cksum` value from the pipeline as soon as possible to
  start `finalize-progress-dialog` using it.
- We need to launch `wl-copy` much later if possible, so that the timing
  of other events will line up where we need them to happen.
- We need to ensure the security of the data from `lpass show`, passed
  in a pipeline to `wl-copy`, which happens to be at the end of that
  same pipeline where `cksum` is called using `tee`.

The solution was to wrap the `lpass show` command in a subshell, and
then wrap the `wl-copy` command in another subshell.  This allows us to
delay the `wl-copy` command by sleeping for 500ms.  This sleep duration
was chosen to ensure that the `wl-paste --watch` script is triggered
AFTER the `rofi` window closes.  This is because the `rofi` window
closes after the `copy-account-field` function completes.  The `disown`
command is used to allow the entire pipeline to run in the background
and the `rofi` window to close as well.  I adjusted the sleep duration
to fit the timing requirement (measured between 367.467521 to 559.041926
milliseconds).