Add security TODO item: Replace `cksum`
The use of cksum was only intended as a proof-of-concept for comparing
some kind of hashed value gathered from `wl-paste --watch` in the
`start-watch-clipboard` forked background process to the hashed value
gathered from the PIPE `tee` sample in the `finalize-progress-dialog`
function. This was done to ensure that the `wl-paste --watch` process
was able to capture the clipboard contents and that the
`finalize-progress-dialog` function was able to capture & validate the
same clipboard contents as well.

The `cksum` command was used because it is widely available on most
Unix-like systems and it is a simple hashing algorithm that is extremely
fast, and easy to understand and use.  However, it is not secure at all
and should not be used in any kind of production code.  It is only used
here as a proof-of-concept to show that the two hashed values are the
same, and that we can validate some sensitive data was actually copied
to the clipboard so the "security screen" progress dialog box can be
safely closed.  The need for passing some value as a command-line
parameter meant that we should not pass the plaintext password to the
forked `rofi-lpass` script when invoked in `finalize-progress-dialog`
mode.  Thus, the need for some hashed value to compare each.

While debugging the timing issues with the `wl-paste --watch` process
and `finalize-progress-dialog` function, the `cksum` command was needed
also to reduce timing impacts from those portions of the code.

The next step is to implement a more secure hashing algorithm to compare
the two hashed values. This will be done in some future commit(s).